+++ .dialyzer_ignore.exs
@@ -2,6 +2,6 @@
 #
 # See https://github.com/jeremyjh/dialyxir#elixir-term-format
 [
-  {"lib/nerves_livebook/fwup.ex", :no_return, 31},
-  {"lib/nerves_livebook/application.ex", :pattern_match, 99}
+  {"lib/nerves_talk/fwup.ex", :no_return, 31},
+  {"lib/nerves_talk/application.ex", :pattern_match, 99}
 ]