+++ config/target.exs
@@ -30,8 +30,8 @@ config :nerves, :erlinit, update_clock: true
 
 config :nerves_ssh,
   daemon_option_overrides: [
-    {:pwdfun, &NervesLivebook.ssh_check_pass/2},
-    {:auth_method_kb_interactive_data, &NervesLivebook.ssh_show_prompt/3}
+    {:pwdfun, &NervesTalk.ssh_check_pass/2},
+    {:auth_method_kb_interactive_data, &NervesTalk.ssh_show_prompt/3}
   ]
 
 config :mdns_lite,