+++ test/nerves_talk/github_release_test.exs
@@ -1,6 +1,6 @@
-defmodule NervesLivebook.GithubReleaseTest do
+defmodule NervesTalk.GithubReleaseTest do
   use ExUnit.Case
-  alias NervesLivebook.GithubRelease
+  alias NervesTalk.GithubRelease
 
   test "version/1" do
     assert "0.1.1" == GithubRelease.version(example_response())
@@ -10,19 +10,19 @@ defmodule NervesLivebook.GithubReleaseTest do
     response = example_response()
 
     assert {:ok,
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi0.fw"} ==
-             GithubRelease.firmware_url(response, "nerves_livebook_rpi0.fw")
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi0.fw"} ==
+             GithubRelease.firmware_url(response, "nerves_talk_rpi0.fw")
 
     assert {:ok,
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_bbb.fw"} ==
-             GithubRelease.firmware_url(response, "nerves_livebook_bbb.fw")
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_bbb.fw"} ==
+             GithubRelease.firmware_url(response, "nerves_talk_bbb.fw")
 
     assert {:ok,
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi4.fw"} ==
-             GithubRelease.firmware_url(response, "nerves_livebook_rpi4.fw")
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi4.fw"} ==
+             GithubRelease.firmware_url(response, "nerves_talk_rpi4.fw")
 
     assert {:error, :not_found} ==
-             NervesLivebook.GithubRelease.firmware_url(response, "nerves_livebook_host.fw")
+             NervesTalk.GithubRelease.firmware_url(response, "nerves_talk_host.fw")
   end
 
   defp example_response() do
@@ -30,13 +30,13 @@ defmodule NervesLivebook.GithubReleaseTest do
       "assets" => [
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_bbb.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_bbb.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_235,
           "label" => "",
-          "name" => "nerves_livebook_bbb.fw",
+          "name" => "nerves_talk_bbb.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjM1",
           "size" => 27_959_165,
           "state" => "uploaded",
@@ -65,17 +65,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708235"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708235"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_bbb.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_bbb.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_230,
           "label" => "",
-          "name" => "nerves_livebook_bbb.zip",
+          "name" => "nerves_talk_bbb.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjMw",
           "size" => 28_268_273,
           "state" => "uploaded",
@@ -104,17 +104,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708230"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708230"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_npi_imx6ull.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_npi_imx6ull.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_226,
           "label" => "",
-          "name" => "nerves_livebook_npi_imx6ull.fw",
+          "name" => "nerves_talk_npi_imx6ull.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI2",
           "size" => 26_792_424,
           "state" => "uploaded",
@@ -143,17 +143,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708226"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708226"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_npi_imx6ull.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_npi_imx6ull.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_231,
           "label" => "",
-          "name" => "nerves_livebook_npi_imx6ull.zip",
+          "name" => "nerves_talk_npi_imx6ull.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjMx",
           "size" => 27_099_401,
           "state" => "uploaded",
@@ -182,17 +182,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708231"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708231"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_osd32mp1.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_osd32mp1.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_221,
           "label" => "",
-          "name" => "nerves_livebook_osd32mp1.fw",
+          "name" => "nerves_talk_osd32mp1.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjIx",
           "size" => 26_788_914,
           "state" => "uploaded",
@@ -221,17 +221,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708221"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708221"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_osd32mp1.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_osd32mp1.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_223,
           "label" => "",
-          "name" => "nerves_livebook_osd32mp1.zip",
+          "name" => "nerves_talk_osd32mp1.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjIz",
           "size" => 32_156_996,
           "state" => "uploaded",
@@ -260,17 +260,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708223"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708223"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_224,
           "label" => "",
-          "name" => "nerves_livebook_rpi.fw",
+          "name" => "nerves_talk_rpi.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI0",
           "size" => 42_968_548,
           "state" => "uploaded",
@@ -299,17 +299,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708224"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708224"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_219,
           "label" => "",
-          "name" => "nerves_livebook_rpi.zip",
+          "name" => "nerves_talk_rpi.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjE5",
           "size" => 43_276_929,
           "state" => "uploaded",
@@ -338,17 +338,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708219"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708219"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi0.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi0.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_227,
           "label" => "",
-          "name" => "nerves_livebook_rpi0.fw",
+          "name" => "nerves_talk_rpi0.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI3",
           "size" => 41_786_592,
           "state" => "uploaded",
@@ -377,17 +377,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708227"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708227"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi0.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi0.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_222,
           "label" => "",
-          "name" => "nerves_livebook_rpi0.zip",
+          "name" => "nerves_talk_rpi0.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjIy",
           "size" => 42_104_612,
           "state" => "uploaded",
@@ -416,17 +416,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708222"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708222"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi2.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi2.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_232,
           "label" => "",
-          "name" => "nerves_livebook_rpi2.fw",
+          "name" => "nerves_talk_rpi2.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjMy",
           "size" => 32_141_641,
           "state" => "uploaded",
@@ -455,17 +455,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708232"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708232"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi2.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi2.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_229,
           "label" => "",
-          "name" => "nerves_livebook_rpi2.zip",
+          "name" => "nerves_talk_rpi2.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI5",
           "size" => 32_477_213,
           "state" => "uploaded",
@@ -494,17 +494,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708229"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708229"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi3.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi3.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_228,
           "label" => "",
-          "name" => "nerves_livebook_rpi3.fw",
+          "name" => "nerves_talk_rpi3.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI4",
           "size" => 42_594_613,
           "state" => "uploaded",
@@ -533,17 +533,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708228"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708228"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi3.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi3.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_225,
           "label" => "",
-          "name" => "nerves_livebook_rpi3.zip",
+          "name" => "nerves_talk_rpi3.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjI1",
           "size" => 42_908_648,
           "state" => "uploaded",
@@ -572,17 +572,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708225"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708225"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi3a.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi3a.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_233,
           "label" => "",
-          "name" => "nerves_livebook_rpi3a.fw",
+          "name" => "nerves_talk_rpi3a.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjMz",
           "size" => 43_110_865,
           "state" => "uploaded",
@@ -611,17 +611,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708233"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708233"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi3a.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi3a.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_220,
           "label" => "",
-          "name" => "nerves_livebook_rpi3a.zip",
+          "name" => "nerves_talk_rpi3a.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjIw",
           "size" => 43_422_806,
           "state" => "uploaded",
@@ -650,17 +650,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708220"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708220"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi4.fw",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi4.fw",
           "content_type" => "application/octet-stream",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_218,
           "label" => "",
-          "name" => "nerves_livebook_rpi4.fw",
+          "name" => "nerves_talk_rpi4.fw",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjE4",
           "size" => 49_039_323,
           "state" => "uploaded",
@@ -689,17 +689,17 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708218"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708218"
         },
         %{
           "browser_download_url" =>
-            "https://github.com/nerves-livebook/nerves_livebook/releases/download/v0.1.1/nerves_livebook_rpi4.zip",
+            "https://github.com/nerves-livebook/nerves_talk/releases/download/v0.1.1/nerves_talk_rpi4.zip",
           "content_type" => "application/zip",
           "created_at" => "2021-04-23T20:18:42Z",
           "download_count" => 0,
           "id" => 35_708_234,
           "label" => "",
-          "name" => "nerves_livebook_rpi4.zip",
+          "name" => "nerves_talk_rpi4.zip",
           "node_id" => "MDEyOlJlbGVhc2VBc3NldDM1NzA4MjM0",
           "size" => 49_407_934,
           "state" => "uploaded",
@@ -728,11 +728,11 @@ defmodule NervesLivebook.GithubReleaseTest do
             "url" => "https://api.github.com/users/nerves-livebook"
           },
           "url" =>
-            "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/assets/35708234"
+            "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/assets/35708234"
         }
       ],
       "assets_url" =>
-        "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/41924363/assets",
+        "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/41924363/assets",
       "author" => %{
         "avatar_url" => "https://avatars.githubusercontent.com/u/64669?v=4",
         "events_url" => "https://api.github.com/users/nerves-livebook/events{/privacy}",
@@ -756,21 +756,19 @@ defmodule NervesLivebook.GithubReleaseTest do
       "body" => "Initial CI-built release",
       "created_at" => "2021-04-23T18:36:35Z",
       "draft" => false,
-      "html_url" => "https://github.com/nerves-livebook/nerves_livebook/releases/tag/v0.1.1",
+      "html_url" => "https://github.com/nerves-livebook/nerves_talk/releases/tag/v0.1.1",
       "id" => 41_924_363,
       "name" => "v0.1.1",
       "node_id" => "MDc6UmVsZWFzZTQxOTI0MzYz",
       "prerelease" => false,
       "published_at" => "2021-04-23T20:19:57Z",
       "tag_name" => "v0.1.1",
-      "tarball_url" =>
-        "https://api.github.com/repos/nerves-livebook/nerves_livebook/tarball/v0.1.1",
+      "tarball_url" => "https://api.github.com/repos/nerves-livebook/nerves_talk/tarball/v0.1.1",
       "target_commitish" => "main",
       "upload_url" =>
-        "https://uploads.github.com/repos/nerves-livebook/nerves_livebook/releases/41924363/assets{?name,label}",
-      "url" => "https://api.github.com/repos/nerves-livebook/nerves_livebook/releases/41924363",
-      "zipball_url" =>
-        "https://api.github.com/repos/nerves-livebook/nerves_livebook/zipball/v0.1.1"
+        "https://uploads.github.com/repos/nerves-livebook/nerves_talk/releases/41924363/assets{?name,label}",
+      "url" => "https://api.github.com/repos/nerves-livebook/nerves_talk/releases/41924363",
+      "zipball_url" => "https://api.github.com/repos/nerves-livebook/nerves_talk/zipball/v0.1.1"
     }
   end
 end