+++ test/nerves_talk_test.exs
@@ -1,13 +1,13 @@
-defmodule NervesLivebookTest do
+defmodule NervesTalkTest do
   use ExUnit.Case
 
   test "check_internet!/0 doesn't raise on host" do
-    NervesLivebook.check_internet!()
+    NervesTalk.check_internet!()
   end
 
   test "version/0" do
     # Just check that it's a parsable version since
     # there's no second source of truth to my knowledge.
-    Version.parse!(NervesLivebook.version())
+    Version.parse!(NervesTalk.version())
   end
 end